+++ src/main/java/net/ultrad00d/ForgottenCantrips/cantrip/EmpowerCantripBuffCantripLogic.java
@@ -1,0 +1,15 @@
+package net.ultrad00d.ForgottenCantrips.cantrip;
+
+import com.mna.api.cantrips.ICantrip;
+
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.player.Player;
+import net.ultrad00d.ForgottenCantrips.registry.EffectRegistry;
+
+public class EmpowerCantripBuffCantripLogic implements IEmpowerCantripLogic, ICantripLogic
+{
+    @Override
+    public void run(Player player, ICantrip cantrip, InteractionHand hand) {
+        applyBuff(player, EffectRegistry.CANTRIP_BUFF);
+    }
+}