+++ src/main/java/net/ultrad00d/ForgottenCantrips/event/EventBusEvents.java
@@ -6,6 +6,7 @@ import net.minecraftforge.fml.common.Mod;
 import net.ultrad00d.ForgottenCantrips.ForgottenCantrips;
 import net.ultrad00d.ForgottenCantrips.entity.OldWizard;
 import net.ultrad00d.ForgottenCantrips.entity.SpectralDonkey;
+import net.ultrad00d.ForgottenCantrips.entity.SpectralSlime;
 import net.ultrad00d.ForgottenCantrips.registry.EntityRegistry;
 
 @Mod.EventBusSubscriber(modid = ForgottenCantrips.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -13,6 +14,7 @@ public class EventBusEvents {
     @SubscribeEvent
     public static void registerAttributes(EntityAttributeCreationEvent event) {
         event.put(EntityRegistry.SPECTRAL_DONKEY.get(), SpectralDonkey.createAttributes());
+        event.put(EntityRegistry.SPECTRAL_SLIME.get(), SpectralSlime.createAttributes().build());
         event.put(EntityRegistry.OLD_WIZARD.get(), OldWizard.createAttributes());
     }
 }