+++ src/main/java/net/ultrad00d/ForgottenCantrips/registry/EntityRegistry.java
@@ -10,6 +10,7 @@ import net.ultrad00d.ForgottenCantrips.ForgottenCantrips;
 import net.ultrad00d.ForgottenCantrips.entity.OldWizard;
 import net.ultrad00d.ForgottenCantrips.entity.SpectralBoat;
 import net.ultrad00d.ForgottenCantrips.entity.SpectralDonkey;
+import net.ultrad00d.ForgottenCantrips.entity.SpectralSlime;
 
 public final class EntityRegistry {
     public static final DeferredRegister<EntityType<?>> ENTITY_TYPES =
@@ -35,6 +36,14 @@ public final class EntityRegistry {
                     EntityType.Builder.of(OldWizard::new, MobCategory.MISC)
                             .sized(0.7F, 1.15F)
                             .build(ForgottenCantrips.MOD_ID + ":old_wizard")
+            );
+
+    public static final RegistryObject<EntityType<SpectralSlime>> SPECTRAL_SLIME =
+            ENTITY_TYPES.register("spectral_slime", () ->
+                    EntityType.Builder.of(SpectralSlime::new, MobCategory.CREATURE)
+                            .sized(0.52F, 0.52F)
+                            .clientTrackingRange(10)
+                            .build("forgotten_cantrips:spectral_slime")
             );
 
     public static void register(IEventBus eventBus) {