+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 - 2026-07-05
+
+### Added
+- Spectral Armor cantrip
+- Villager Mind-Wiping cantrip
+
+### Changed
+- Force Consume cantrips was renamed to Devour
+
+### Fixed
+- Devour cantrip lightning effect fixed
+
+### Changed
+
+Memory Emeralds can now only be applied using the Villager Mind-Wiper cantrip
+
 ## 0.3.0 - 2026-06-28
 
 ### Added